+++ src/Pages/RecipeFinder.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useQuery } from '@apollo/react-hooks';
 import { gql } from 'graphql-tag';
-import { Typography, Box, Container, Button, TextField, FormControl, InputLabel, Select, MenuItem} from '@mui/material';
+import { Typography, Box, Button, TextField, FormControl, InputLabel, Select, MenuItem} from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
 
 const getRecipes = gql`
